rename fpma api service
