+++ src/app/app.module.ts
@@ -15,7 +15,7 @@ import { ActualityPage } from '../pages/actuality/actuality';
 import { StkNewsPage } from '../pages/stk-news/stk-news';
 import { MainTabsPage } from '../pages/main-tabs/main-tabs';
 import { PresentationPage } from '../pages/presentation/presentation';
-import { FpmaApiService } from '../services/fpma-api/fpma-api-service';
+import { FpmaApiService } from '../services/fpma-api/fpma-api.service';
 
 @NgModule({
   declarations: [
